Train Reservation System in C language.
A train reservation system project in C language is a program designed to manage the reservation and cancellation of train tickets. It allows users to search for available train routes, choose their preferred travel dates and times, and reserve seats on a train. The program  also have the capability to display information about train schedules, seat availability, and fares.